Corection methode min et Création methode melanger

+++ src/fr/ecole3il/rodez2023/perlin/math/BruitAleatoire.java
@@ -18,6 +18,6 @@ public class BruitAleatoire extends Bruit2D {
     public double bruit2D(double x, double y)
     {
         Random random = new Random();
-        return (double)random.nextInt(100);
+        return random.nextDouble(100);
     }
 }
